major model load refactor
