+++ modules/processing_info.py
@@ -41,11 +41,12 @@ def create_infotext(p: StableDiffusionProcessing, all_prompts=None, all_seeds=No
         # basic
         "Steps": p.steps,
         "Seed": all_seeds[index],
-        "Sampler": p.sampler_name,
-        "CFG scale": p.cfg_scale,
+        "Sampler": p.sampler_name if p.sampler_name != 'Default' else None,
+        "CFG scale": p.cfg_scale if p.cfg_scale > 1.0 else None,
+        "CFG end": p.cfg_end if p.cfg_end < 1.0 else None,
         "Size": f"{p.width}x{p.height}" if hasattr(p, 'width') and hasattr(p, 'height') else None,
         "Batch": f'{p.n_iter}x{p.batch_size}' if p.n_iter > 1 or p.batch_size > 1 else None,
-        "Parser": shared.opts.prompt_attention,
+        "Parser": shared.opts.prompt_attention.split()[0],
         "Model": None if (not shared.opts.add_model_name_to_info) or (not shared.sd_model.sd_checkpoint_info.model_name) else shared.sd_model.sd_checkpoint_info.model_name.replace(',', '').replace(':', ''),
         "Model hash": getattr(p, 'sd_model_hash', None if (not shared.opts.add_model_hash_to_info) or (not shared.sd_model.sd_model_hash) else shared.sd_model.sd_model_hash),
         "VAE": (None if not shared.opts.add_model_name_to_info or sd_vae.loaded_vae_file is None else os.path.splitext(os.path.basename(sd_vae.loaded_vae_file))[0]) if p.full_quality else 'TAESD',
