+++ scripts/xyz_grid_classes.py
@@ -99,7 +99,7 @@ axis_options = [
     AxisOption("[Param] Height", int, apply_field("height")),
     AxisOption("[Param] Seed", int, apply_seed),
     AxisOption("[Param] Steps", int, apply_field("steps")),
-    AxisOption("[Param] CFG scale", float, apply_field("cfg_scale")),
+    AxisOption("[Param] Guidance scale", float, apply_field("cfg_scale")),
     AxisOption("[Param] Guidance end", float, apply_field("cfg_end")),
     AxisOption("[Param] Variation seed", int, apply_field("subseed")),
     AxisOption("[Param] Variation strength", float, apply_field("subseed_strength")),
@@ -125,7 +125,7 @@ axis_options = [
     AxisOption("[Refine] Sampler", str, apply_hr_sampler_name, fmt=format_value, confirm=confirm_samplers, choices=lambda: [x.name for x in sd_samplers.samplers]),
     AxisOption("[Refine] Denoising strength", float, apply_field("denoising_strength")),
     AxisOption("[Refine] Hires steps", int, apply_field("hr_second_pass_steps")),
-    AxisOption("[Refine] CFG scale", float, apply_field("image_cfg_scale")),
+    AxisOption("[Refine] Guidance scale", float, apply_field("image_cfg_scale")),
     AxisOption("[Refine] Guidance rescale", float, apply_field("diffusers_guidance_rescale")),
     AxisOption("[Refine] Refiner start", float, apply_field("refiner_start")),
     AxisOption("[Refine] Refiner steps", float, apply_field("refiner_steps")),
